+++ tests/test_codegen_pipeline.py
@@ -17,13 +17,24 @@
 import contextlib
 import copy
 import io
+import json
 import sys
 import tempfile
 import unittest
 from pathlib import Path
 from unittest import mock
 
+import postprocess_models
 import preprocess_schemas
+
+try:
+    from pydantic import ValidationError
+
+    from ucp_sdk.models.schemas.shopping.types.description import Description
+
+    HAVE_SDK = True
+except ImportError:  # pragma: no cover
+    HAVE_SDK = False
 
 
 class SchemaNormalizationTest(unittest.TestCase):
@@ -533,6 +544,97 @@ class PipelineDependencyTest(unittest.TestCase):
         )
         self.assertEqual(set(parent_variant["required"]), {"id", "child"})
         self.assertEqual(child_variant["required"], ["value"])
+
+
+@unittest.skipUnless(
+    HAVE_SDK, "requires the installed package (pip install -e .)"
+)
+class DescriptionMinPropertiesTest(unittest.TestCase):
+    """description.json declares minProperties: 1 at the schema root."""
+
+    def test_empty_instance_rejected(self):
+        with self.assertRaisesRegex(ValidationError, "[Aa]t least 1"):
+            Description()
+
+    def test_empty_mapping_rejected(self):
+        with self.assertRaisesRegex(ValidationError, "[Aa]t least 1"):
+            Description.model_validate({})
+
+    def test_single_declared_field_accepted(self):
+        self.assertEqual(Description(plain="hello").plain, "hello")
+
+    def test_explicit_null_key_counts_as_present(self):
+        # {"html": null} has one property per JSON Schema's key counting.
+        Description.model_validate({"html": None})
+
+    def test_extra_field_counts_as_present(self):
+        # extra="allow": an unknown key is a present property.
+        Description.model_validate({"x-vendor-note": "hi"})
+
+    def test_all_fields_accepted(self):
+        Description(plain="p", html="<p>p</p>", markdown="p")
+
+
+class InjectorTest(unittest.TestCase):
+    """The post-generation injector's own behavior."""
+
+    SCHEMA = {
+        "title": "Sample",
+        "type": "object",
+        "minProperties": 2,
+        "properties": {"a": {"type": "string"}, "b": {"type": "string"}},
+    }
+
+    MODULE = (
+        "from __future__ import annotations\n"
+        "\n"
+        "from pydantic import BaseModel, ConfigDict\n"
+        "\n"
+        "\n"
+        "class Sample(BaseModel):\n"
+        '    """A sample."""\n'
+        "\n"
+        "    model_config = ConfigDict(\n"
+        '        extra="allow",\n'
+        "    )\n"
+        "    a: str | None = None\n"
+        "    b: str | None = None\n"
+    )
+
+    def test_injects_validator_with_declared_minimum(self):
+        out = postprocess_models.inject_min_properties(self.MODULE, "Sample", 2)
+        self.assertIn("model_validator", out)
+        self.assertIn("at least 2", out.lower())
+
+    @unittest.skipUnless(HAVE_SDK, "executing the module needs pydantic")
+    def test_injected_validator_enforces_count(self):
+        out = postprocess_models.inject_min_properties(self.MODULE, "Sample", 2)
+        namespace: dict = {}
+        exec(compile(out, "<injected>", "exec"), namespace)  # noqa: S102
+        sample_cls = namespace["Sample"]
+        with self.assertRaises(ValidationError):
+            sample_cls(a="only-one")
+        sample_cls(a="one", b="two")
+
+    def test_injection_is_idempotent(self):
+        once = postprocess_models.inject_min_properties(
+            self.MODULE, "Sample", 2
+        )
+        twice = postprocess_models.inject_min_properties(once, "Sample", 2)
+        self.assertEqual(once, twice)
+
+    def test_schema_scan_finds_root_constraints(self):
+        with tempfile.TemporaryDirectory() as tmp:
+            sub = Path(tmp) / "sub"
+            sub.mkdir()
+            (sub / "sample.json").write_text(json.dumps(self.SCHEMA))
+            (sub / "plain.json").write_text(
+                json.dumps(
+                    {"title": "Plain", "type": "object", "properties": {}}
+                )
+            )
+            found = postprocess_models.find_root_min_properties(Path(tmp))
+        self.assertEqual(found, {"Sample": 2})
 
 
 if __name__ == "__main__":
